chore(select): forward refs to select dom element
