+++ __tests__/renderer/shared/components/Forms/LabeledSelect/index.test.js
@@ -1,32 +1,16 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import LabeledSelect from 'shared/components/Forms/LabeledSelect';
-import Label from 'shared/components/Forms/Label';
-import Select from 'shared/components/Forms/Select';
+import LabeledSelectContainer from 'shared/components/Forms/LabeledSelect';
+import LabeledSelect from 'shared/components/Forms/LabeledSelect/LabeledSelect';
 
 const mountContainer = (props = {}) => {
-  return mount(<LabeledSelect {...props} />);
+  return mount(<LabeledSelectContainer {...props} />);
 };
 
 describe('<LabeledSelect />', () => {
-  it('renders a label', () => {
-    const wrapper = mountContainer({ id: 'currency', label: 'Currency' });
-    const label = wrapper.find(Label);
-    expect(label.exists()).toBe(true);
-    expect(label.props()).toEqual(expect.objectContaining({ label: 'Currency', htmlFor: 'currency' }));
-  });
-
-  it('renders a select', () => {
-    const children = (
-      <React.Fragment>
-        <option value="USD">United States Dollar</option>
-        <option value="EUR">Euro</option>
-      </React.Fragment>
-    );
-    const wrapper = mountContainer({ id: 'currency', label: 'Currency', children });
-    const select = wrapper.find(Select);
-    expect(select.exists()).toBe(true);
-    expect(select.props()).toEqual(expect.objectContaining({ id: 'currency', children }));
+  it('forwards the ref to the component', () => {
+    const wrapper = mountContainer({ ref: React.createRef() });
+    expect(wrapper).toForwardRefTo(LabeledSelect);
   });
 });
