+++ src/renderer/shared/components/Forms/LabeledSelect/index.js
@@ -1,1 +1,5 @@
-export { default } from './LabeledSelect';
+import withForwardedRef from 'shared/hocs/withForwardedRef';
+
+import LabeledSelect from './LabeledSelect';
+
+export default withForwardedRef()(LabeledSelect);
