+++ src/renderer/shared/components/Forms/Select/Select.js
@@ -1,24 +1,28 @@
 import React from 'react';
 import classNames from 'classnames';
-import { string } from 'prop-types';
+import { string, func } from 'prop-types';
 
 import styles from './Select.scss';
 
 export default class Select extends React.PureComponent {
   static propTypes = {
-    className: string
+    className: string,
+    forwardedRef: func
   };
 
   static defaultProps = {
-    className: null
+    className: null,
+    forwardedRef: null
   };
 
   render() {
-    const { className, ...passDownProps } = this.props;
+    const { className, forwardedRef, ...passDownProps } = this.props;
 
     return (
+
       <select
         {...passDownProps}
+        ref={forwardedRef}
         className={classNames(styles.select, className)}
       />
     );
